Revert "Modularize game elements"
