+++ src/components/game/trainSystem.ts
@@ -169,26 +169,26 @@ export const MAX_TRAINS_MOBILE = 8; // Fewer trains on mobile for performance
 export const TRAIN_SPAWN_INTERVAL = 1.5;
 export const TRAIN_SPAWN_INTERVAL_MOBILE = 4.0; // Slower spawning on mobile
 
-/** Station stop duration in seconds - internal use */
-const STATION_STOP_DURATION = 2.0;
+/** Station stop duration in seconds */
+export const STATION_STOP_DURATION = 2.0;
 
-/** Train maximum age in seconds - internal use */
-const TRAIN_MAX_AGE = 120;
+/** Train maximum age in seconds */
+export const TRAIN_MAX_AGE = 120;
 
-/** Safe following distance in tile units (triggers slowdown) - internal use */
+/** Safe following distance in tile units (triggers slowdown) */
 const SAFE_FOLLOWING_DISTANCE = 0.8;
 
-/** Minimum speed when slowing for train ahead (never fully stop to prevent deadlocks) - internal use */
+/** Minimum speed when slowing for train ahead (never fully stop to prevent deadlocks) */
 const MIN_FOLLOWING_SPEED = 0.3;
 
-/** Carriage count for train types - internal use */
-const TRAIN_CARRIAGE_COUNTS = {
+/** Carriage count for train types */
+export const TRAIN_CARRIAGE_COUNTS = {
   passenger: { min: 5, max: 8 },
   freight: { min: 6, max: 10 },
 };
 
-/** Carriage spacing (in tile progress units) - spacing between carriages - internal use */
-const CARRIAGE_SPACING = 0.28;
+/** Carriage spacing (in tile progress units) - spacing between carriages */
+export const CARRIAGE_SPACING = 0.28;
 
 // ============================================================================
 // Train Creation Functions
