--- src/core/types/index.ts
@@ -1,3 +1,0 @@
-// Core types index - re-export all core types
-export * from './grid';
-export * from './rendering';
