--- src/games/isocity/index.ts
@@ -1,4 +1,0 @@
-// IsoCity game module index
-// Re-exports all IsoCity-specific types and utilities
-
-export * from './types';
