--- src/games/isocity/types/economy.ts
@@ -1,47 +1,0 @@
-// IsoCity economy and statistics type definitions
-
-export interface Stats {
-  population: number;
-  jobs: number;
-  money: number;
-  income: number;
-  expenses: number;
-  happiness: number;
-  health: number;
-  education: number;
-  safety: number;
-  environment: number;
-  demand: {
-    residential: number;
-    commercial: number;
-    industrial: number;
-  };
-}
-
-export interface BudgetCategory {
-  name: string;
-  funding: number;
-  cost: number;
-}
-
-export interface Budget {
-  police: BudgetCategory;
-  fire: BudgetCategory;
-  health: BudgetCategory;
-  education: BudgetCategory;
-  transportation: BudgetCategory;
-  parks: BudgetCategory;
-  power: BudgetCategory;
-  water: BudgetCategory;
-}
-
-// Cached economy data for a city
-export interface CityEconomy {
-  population: number;
-  jobs: number;
-  income: number;
-  expenses: number;
-  happiness: number;
-  // Timestamp of last calculation for cache invalidation
-  lastCalculated: number;
-}
