--- src/games/isocity/types/index.ts
@@ -1,7 +1,0 @@
-// IsoCity types index - re-export all IsoCity-specific types
-export * from './buildings';
-export * from './zones';
-export * from './economy';
-export * from './services';
-export * from './game';
-export * from './entities';
