--- src/games/isocity/types/zones.ts
@@ -1,3 +1,0 @@
-// IsoCity zone type definitions
-
-export type ZoneType = 'none' | 'residential' | 'commercial' | 'industrial';
